requirements.txt
streamlit
opencv-python
pytesseract
numpy
Pillow
imutils
transformers
torch
